// app-tabman-pwa-vite/src/types/rdo-types.ts
export interface RdoMotivo {
  RDOMOTIVOCOD: number;
  DESCRICAO: string;
  SIGLA: string;
  ATIVO: 'S' | 'N';
  PRODUTIVO: 'S' | 'N';
  TOLERANCIA: number | null;
  PENALIDADE: number | null;
  WTCATEGORIA: string | null;
}